udma: support jfs flush.
category: feature
bugzilla: https://gitee.com/openeuler/umdk/issues/ID45DE
CVE: NA

-----------------------------------------------------------

In the IO process, if the sq queue encounters an anomaly,
the hardware will mark the queue as faulty; in certain
scenarios, the software may also proactively switch the
queue status and halt operations. In both of these
situations, it is necessary to clear the queue and
generate a Flush CQE for any incomplete SQEs. The
functionality of the Flush CQE is accomplished
through the collaboration of both software and hardware.

Signed-off-by: Jiahao Zhang <zhangjiahao14@huawei.com>