udma: support for querying the root cause of AE errors
category: feature
bugzilla: https://gitee.com/openeuler/umdk/issues/ID45DE
CVE: NA

-----------------------------------------------------------

Class B Fault: The characteristic of this type of error is
that the physical topology of the network is intact, but
the fault cannot be attributed to a single transmission
task and requires handling of a series of related tasks.
This commonly occurs when a single Jetty, a single Entity,
or a single TP-level resource fails. It is necessary to
perform relevant processing on all resources of the faulty
unit as a whole. When a Class B fault occurs, an AE
(Asynchronous Event) event is reported.
1.support users in invoking the user_ctl interface to query
the root cause of errors when retrieving AE(Asynchronous Event)
information.

Signed-off-by: Jiahao Zhang <zhangjiahao14@huawei.com>